Updated several migrations to avoid model references.
Based on readings I found on the web, database migrations that require more than a few steps should be avoided. The best approach for these scenarios is to use the schema.rb dump to replace your old structure and then manually insert data as required.

If your database is version 102 or older, you will need to use the schema.rb to redefine your database instead of submitting a rake db:migrate call.

git-svn-id: https://svn.csc.ncsu.edu:8443/svn/research-expertiza/trunk@638 392d7b7b-3f31-0410-9dc3-c95a9635ea79

+++ db/migrate/114_update_controllers_for_response.rb
@@ -3,28 +3,40 @@ class UpdateControllersForResponse < ActiveRecord::Migration
     perm = Permission.find_by_name("do assignments")
     
     controller = SiteController.find_or_create_by_name("response")
-    controller.permission_id = perm.id
-    controller.save    
+    if controller
+      controller.permission_id = perm.id
+      controller.save
+    end
     
     controller = SiteController.find_by_name("review")
-    ControllerAction.find_all_by_site_controller_id(controller.id).each{|action| action.destroy}
-    controller.destroy
+    if controller 
+      ControllerAction.find_all_by_site_controller_id(controller.id).each{|action| action.destroy}
+      controller.destroy
+    end
     
     controller = SiteController.find_by_name("reviewing")
-    ControllerAction.find_all_by_site_controller_id(controller.id).each{|action| action.destroy}
-    controller.destroy
+    if controller
+      ControllerAction.find_all_by_site_controller_id(controller.id).each{|action| action.destroy}
+      controller.destroy
+    end
     
     controller = SiteController.find_by_name("review_feedback")
-    ControllerAction.find_all_by_site_controller_id(controller.id).each{|action| action.destroy}
-    controller.destroy
+    if controller
+      ControllerAction.find_all_by_site_controller_id(controller.id).each{|action| action.destroy}
+      controller.destroy
+    end
     
     controller = SiteController.find_by_name("review_of_review")
-    ControllerAction.find_all_by_site_controller_id(controller.id).each{|action| action.destroy}
-    controller.destroy    
+    if controller
+      ControllerAction.find_all_by_site_controller_id(controller.id).each{|action| action.destroy}
+      controller.destroy    
+    end
     
     controller = SiteController.find_by_name("teammate_review")
-    ControllerAction.find_all_by_site_controller_id(controller.id).each{|action| action.destroy}
-    controller.destroy      
+    if controller
+      ControllerAction.find_all_by_site_controller_id(controller.id).each{|action| action.destroy}
+      controller.destroy
+    end
     
     Role.rebuild_cache
   end
